+++ src/db/mysql/MySqlAPI.h
@@ -207,7 +207,7 @@ public:
         std::string service_name);
 
     /// Update the state of a transfer inside a session reuse job
-    virtual unsigned int updateFileStatusReuse(const TransferFile &file, const std::string &status);
+    virtual long updateFileStatusReuse(const TransferFile &file, const std::string &status);
 
     /// Puts into requestIDs, jobs that have been cancelled, and for which the running fts_url_copy must be killed
     virtual void getCancelJob(std::vector<int>& requestIDs);
